requirements.txt
langchain
langchain-openai
langchain-groq
langchain-community
langsmith
openai
groq
huggingface-hub
python-dotenv
